HACK: Avoid switching out of realtime by removing calls to logger.log()
The controller_manager got stuck when loading rewd_controllers and system throws the warning
'Switched out of RealTime'. Looking at var/log/syslog it appears to switch out when calling the logger.log() method.
Commenting out these calls seems to fix the issue. But it still remains a mystery as to why this is happening.
A possible issue could be updates in system-level libraries. Ethan: The only thing the logger does is push to an
std::queue, so it is possible that std::queue is no longer real-time safe.